Fixed the main view so that it calculates based on the categories. Added editing the database for possible points, category points, and category names.

+++ trunk/web/OSBLE/Models/Courses/Category.cs
@@ -34,12 +34,12 @@ namespace OSBLE.Models.Courses
 
         [Required]
         [Display(Name = "Weight")]
-        public int Points { get; set; }
+        public double Points { get; set; }
 
         /// <summary>
         /// Grading points possible relative to sibling categories in the parent course.
         /// </summary>
-        public int PointsPossible { get; set; }
+        public double PointsPossible { get; set; }
 
         /// <summary>
         /// Used for visual ordering of various categories (tabs in the gradebook).  Defaults to a value of 0.
